refactored backend/subscription responsibilities
