--- microservices/apps/backend/src/application/use-cases/subscription/confirmSubscriptionUseCase.js
@@ -1,14 +1,0 @@
-class ConfirmSubscriptionUseCase
-{
-    constructor(subscriptionService)
-    {
-        this.subscriptionService = subscriptionService;
-    }
-
-    async confirm(token)
-    {
-        return this.subscriptionService.confirmSubscription(token);
-    }
-}
-
-module.exports = ConfirmSubscriptionUseCase;
