--- microservices/apps/backend/src/application/use-cases/subscription/unsubscribeUserUseCase.js
@@ -1,15 +1,0 @@
-class UnsubscribeUserUseCase
-{
-    constructor(subscriptionService)
-    {
-        this.subscriptionService = subscriptionService;
-    }
-
-    async unsubscribe(token)
-    {
-        return this.subscriptionService.unsubscribeUser(token);
-    }
-}
-
-module.exports = UnsubscribeUserUseCase;
-
