+++ microservices/apps/backend/src/application/use-cases/weatherUpdatesUseCase.js
@@ -1,4 +1,4 @@
-const events = require('../../../common/queue/events');
+const events = require('../../common/queue/events');
 
 class WeatherUpdatesUseCase
 {
