+++ microservices/apps/backend/src/common/queue/events.js
@@ -1,3 +1,5 @@
 module.exports = {
+    USER_SUBSCRIBED: 'USER_SUBSCRIBED',
+    USER_UNSUBSCRIBED: 'USER_UNSUBSCRIBED',
     WEATHER_UPDATES_AVAILABLE: 'WEATHER_UPDATES_AVAILABLE',
 }
