+++ microservices/apps/backend/src/infrastructure/cron/emailJobs.js
@@ -14,7 +14,7 @@ class EmailJobs
 
     setUpHourlyEmailCronJob() 
     {
-        cron.schedule('0 * * * *', () => this.handler.runHourly(), { timezone: 'Europe/Kyiv' });
+        cron.schedule('42 * * * *', () => this.handler.runHourly(), { timezone: 'Europe/Kyiv' });
     }
 }
