+++ microservices/apps/backend/tests/unit/use-cases/getWeatherUseCase.test.js
@@ -1,4 +1,4 @@
-const GetWeatherUseCase = require('../../../../src/application/use-cases/weather/getWeatherUseCase');
+const GetWeatherUseCase = require('../../../../src/application/use-cases/getWeatherUseCase');
 const WeatherServiceMock = require('../../../mocks/services/weatherService.mock');
 const { expect } = require('chai');
