--- microservices/apps/subscription/src/common/config/index.js
@@ -1,14 +1,0 @@
-require('dotenv').config({
-    path: `.env.${process.env.NODE_ENV || 'development'}`
-});
-
-const queueUrl = process.env.QUEUE_URL;
-const queueName = process.env.QUEUE_NAME;
-
-module.exports = {
-    queue :
-        {
-            queueUrl,
-            queueName
-        }
-};
