--- microservices/apps/subscription/src/common/queue/IQueuePublisher.js
@@ -1,7 +1,0 @@
-class IQueuePublisher {
-    async publish(type, payload, queue) {
-        throw new Error('Not implemented');
-    }
-}
-
-module.exports = IQueuePublisher;
