--- microservices/apps/subscription/src/common/queue/events.js
@@ -1,4 +1,0 @@
-module.exports = {
-    USER_SUBSCRIBED: 'USER_SUBSCRIBED',
-    USER_UNSUBSCRIBED: 'USER_UNSUBSCRIBED',
-};
